refactor(lint): rename 'eslint' subgen to 'lint'
