+++ generators/index.js
@@ -4,6 +4,6 @@ const Generators = require('yeoman-generator');
 
 module.exports = Generators.Base.extend({
   initializing : function () {
-    this.composeWith('lob:eslint');
+    this.composeWith('lob:lint');
   }
 });
