+++ generators/lint/templates/eslintignore
