+++ generators/lint/templates/eslintrc
