+++ index.js
@@ -1,5 +1,5 @@
 'use strict';
 
 module.exports = {
-  eslint: require('./generators/eslint')
+  lint: require('./generators/lint')
 };
